% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{gix}
\alias{gix}
\alias{gix,character,character-method}
\title{Search a git repository}
\usage{
gix(path, query)

\S4method{gix}{character,character}(path, query)
}
\arguments{
\item{path}{Path to the repository}

\item{query}{The search query}
}
\value{
:TODO:DOCUMENTATION:
}
\description{
Search a git repository
}
\keyword{methods}

