% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{gixit}
\alias{gixit}
\alias{gixit,character-method}
\title{Index a git repository}
\usage{
gixit(path, ...)

\S4method{gixit}{character}(path, callbacks)
}
\arguments{
\item{path}{Path to the repository}

\item{callbacks}{A list of callback functions to enable more
control over how blob's are indexed. Each list item, named by file
extension, is a function with the two arguments 'content' and
'sha'.}
}
\value{
invisible(NULL)
}
\description{
Index a git repository
}
\keyword{methods}

