% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{gix}
\alias{gix}
\alias{gix,git_repository,character-method}
\title{Search a git repository}
\usage{
gix(repo, query)

\S4method{gix}{git_repository,character}(repo, query)
}
\arguments{
\item{repo}{The repository}

\item{query}{The search query}
}
\value{
:TODO:DOCUMENTATION:
}
\description{
Search a git repository
}
\keyword{methods}

